requirements.txt
PIL
paho-mqtt
inky
